### 2018-09-12 #939 @whedon
[PRE REVIEW]: Diff_classifier: Parallelization of multi-particle tracking video analyses
**Submitting author:** @ccurtis7 (<a href="http://orcid.org/0000-0001-6312-392X">Chad Curtis</a>)
**Repository:** <a href="https://github.com/ccurtis7/diff_classifier" target ="_blank">https://github.com/ccurtis7/diff_classifier</a>
**Version:** v0.1-alpha
