### 2018-09-12 05fb8b3
Attempt at fixing paper generation
