+++ paper.bib
@@ -41,8 +41,9 @@
          }
 
 @article{zenodo,
-         title = {Diff_classifier},
+         title = {diff_classifier},
          author = {Curtis C, Rokem A, and Nance E},
          publisher = {Zenodo},
          year = {2018},
+         doi= {10.5281/zenodo.1413740}
          }
