+++ paper.md
@@ -1,51 +1,51 @@
 ---
-title: 'diff_classifier: Parallelization of multi-particle tracking videos'
+title: 'diff_classifier: Parallelization of multi-particle tracking video analyses'
 tags:
-  - Python
-  - multi-particle tracking
-  - Amazon Web Services
-  - parallelization
+- Python
+- multi-particle tracking
+- Amazon Web Services
+- parallelization
 authors:
-  - name: Chad Curtis
-    orcid: 0000-0001-6312-392X
-    affiliation: "1"
-  - name: Ariel Rokem
-    orcid: 0000-0003-0679-1985
-    affiliation: "2"
-  - name: Elizabeth Nance
-    orcid: 0000-0001-7167-7068
-    affiliation: "1"
+- name: Chad Curtis
+  orcid: 0000-0001-6312-392X
+  affiliation: "1"
+- name: Ariel Rokem
+  orcid: 0000-0003-0679-1985
+  affiliation: "2"
+- name: Elizabeth Nance
+  orcid: 0000-0001-7167-7068
+  affiliation: "1"
 affiliations:
- - name: Department of Chemical Engineering, University of Washington
-   index: 1
- - name: eScience Institute, University of Washington
-   index: 2
+- name: Department of Chemical Engineering, University of Washington
+  index: 1
+- name: eScience Institute, University of Washington
+  index: 2
 date: 20 August 2018
 bibliography: paper.bib
 ---
 
 # Summary
 
-The ``diff_classifier`` package seeks to address the issue of scale-up in
-multi-particle tracking (MPT) analyses via a parallelization approach. MPT is a
-powerful analytical tool that has been used in fields ranging from aeronautics
-to oceanography to biomedical engineering [@Pulford:2005]. While a variety of
-tracking algorithms are available to researchers [@Chenouard:2014], a common
-problem is the large time sink in data analysis due to a widespread use of
-graphical user interfaces and the reliance on human input for accurate tracking.
-For example, particle detection often relies on the selection of a quality
-threshold, a numerical quantity distinguishing between “real” particles and
-“fake” particles [@Tineves:2017]. If this threshold is too high, the high number
-of false positives can result in false trajectories that skew results, and in
-extreme cases, cause the code to crash due to a lack of convergence in the
-particle linking step. If the threshold is too low, trajectories will be cut
-short resulting in a bias towards short fast-moving trajectories and could
-result in empty datasets [@Wang:2015].
+The [diff_classifier](https://github.com/ccurtis7/diff_classifier) package seeks
+to address the issue of scale-up in multi-particle tracking (MPT) analyses via a
+parallelization approach. MPT is a powerful analytical tool that has been used
+in fields ranging from aeronautics to oceanography to biomedical engineering
+[@Pulford:2005]. While a variety of tracking algorithms are available to
+researchers [@Chenouard:2014], a common problem is the large time sink in data
+analysis due to a widespread use of graphical user interfaces and the reliance
+on human input for accurate tracking. For example, particle detection often
+relies on the selection of a quality threshold, a numerical quantity
+distinguishing between “real” particles and “fake” particles [@Tineves:2017]. If
+this threshold is too high, the high number of false positives can result in
+false trajectories that skew results, and in extreme cases, cause the code to
+crash due to a lack of convergence in the particle linking step. If the
+threshold is too low, trajectories will be cut short resulting in a bias towards
+short fast-moving trajectories and could result in empty datasets [@Wang:2015].
 
 Due to variations in experimental conditions and image quality, user-selected
 tracking parameters can vary widely from video to video. As parameter selection
 can also vary from user to user, this also brings up the issue of
-reproducibility. ``Diff_classifier`` addresses these issues with regression
+reproducibility. diff_classifier addresses these issues with regression
[…]
