### 2018-09-12 2f5ec02
Removed abstracts from paper.bib
