+++ paper.bib
@@ -5,19 +5,16 @@
          volume = {152},
          number = {5},
          year = {2005},
-         keywords = {target tracking, computational complexity, probability, filtering theory, reviews, estimation theory},
-         abstract = {A concise summary of techniques for multiple target tracking is provided and their main characteristics assessed qualitatively. The techniques have been catergorised into more than 35 different algorithmic types. A comparison chart is provided that lists each algorithm and categorises the processing scheme, data association mechanism, complexity scaling (with number of targets and with state dimension), overall complexity and a subjective performance figure. Although some recent filtering theory developments have been omitted, the survey should serve to demonstrate the large variety of 'classical' estimation theoretic algorithms already available for the design of multiple target tracking systems. A number of areas deserving of further study are identified in the concluding remarks.}
          }
 
 @article{Chenouard:2014,
-        author = {Chenouard N},
-        title = {Objective comparison of particle tracking methods},
-        journal = {Nature Methods},
-        volume = {11},
-        number = {3},
-        year = {2014},
-        abstract = {Particle tracking is of key importance for quantitative analysis of intracellular dynamic processes from time-lapse microscopy image data. Because manually detecting and following large numbers of individual particles is not feasible, automated computational methods have been developed for these tasks by many groups. Aiming to perform an objective comparison of methods, we gathered the community and organized an open competition in which participating teams applied their own methods independently to a commonly defined data set including diverse scenarios. Performance was assessed using commonly defined measures. Although no single method performed best across all scenarios, the results revealed clear differences between the various approaches, leading to notable practical conclusions for users and developers.}
-        }
+         author = {Chenouard N},
+         title = {Objective comparison of particle tracking methods},
+         journal = {Nature Methods},
+         volume = {11},
+         number = {3},
+         year = {2014},
+         }
 
 @article{Tineves:2017,
          author = {Tineves J, Perry N, Schindelin J, Hoopes G M, Reynolds G D, Laplantine E, Bednarek S Y, Shorte S L, and Eliceiri K W},
@@ -25,8 +22,6 @@
          journal = {Methods},
          volume = {115},
          year = {2017},
-         keywords = {Clathin-mediated endocytosis, image analysis, microscopy, open-source software, phototoxicity, single-particle tracking},
[…]
