### 2018-09-12 #939 @ccurtis7
It seems that the test PDF is failing to generate. I looked at a few [examples](https://github.com/openjournals/joss-reviews/issues/485) from other JOSS pre-reviews to see if I could track down a solution. I think it has to do with a stray character or figure that is being interpreted incorrectly. I'm not sure exactly what it is, but I suspect it is somewhere in my paper.bib file. I tried removing the abstracts from the paper.bib file to see if I could resolve the error in a new branch [joss_review](https://github.com/ccurtis7/diff_classifier/tree/joss_review).
