### 2018-09-12 b6cc0fa
Reformmated paper.bib
