+++ paper.bib
@@ -1,42 +1,42 @@
-@article{Pulford:2005,
-         author = {Pulford G W},
-         title = {Taxonomy of multiple target tracking methods},
-         journal = {Iee Proceedings-Radar Sonar and Navigation},
-         volume = {152},
-         number = {5},
-         year = {2005},
-         }
+@ARTICLE{Pulford:2005,
+  author = "Pulford G W",
+  title = "Taxonomy of multiple target tracking methods",
+  journal = "Iee Proceedings-Radar Sonar and Navigation",
+  volume = 152,
+  number = 5,
+  year   = 2005
+}
 
-@article{Chenouard:2014,
-         author = {Chenouard N},
-         title = {Objective comparison of particle tracking methods},
-         journal = {Nature Methods},
-         volume = {11},
-         number = {3},
-         year = {2014},
-         }
+@ARTICLE{Chenouard:2014,
+  author  = "Chenouard N",
+  title   = "Objective comparison of particle tracking methods",
+  journal = "Nature Methods",
+  volume  = 11,
+  number  = 3,
+  year    = 2014
+}
 
-@article{Tineves:2017,
-         author = {Tineves J, Perry N, Schindelin J, Hoopes G M, Reynolds G D, Laplantine E, Bednarek S Y, Shorte S L, and Eliceiri K W},
-         title = {TrackMate: An open and extensible platform for single-particle tracking},
-         journal = {Methods},
-         volume = {115},
-         year = {2017},
-         }
+@ARTICLE{Tineves:2017,
+  author  = "Tineves J, Perry N, Schindelin J, Hoopes G M, Reynolds G D, Laplantine E, Bednarek S Y, Shorte S L, and Eliceiri K W",
+  title   = "TrackMate: An open and extensible platform for single-particle tracking",
+  journal = "Methods",
+  volume  = 115,
+  year    = 2017
+}
 
-@article{Wang:2015,
-         author = {Wang Y Y, Nunn K L, Harit D, McKinley S A, and Lai S K},
-         title = {Minimizing biases associated with tracking analysis of submicron particles in heterogeneous biological fluids},
-         journal = {Journal of Controlled Release},
-         volume = {220},
-         number = {PtA},
-         year = {2015},
-         }
+@ARTICLE{Wang:2015,
+  author  = "Wang Y Y, Nunn K L, Harit D, McKinley S A, and Lai S K",
+  title   = "Minimizing biases associated with tracking analysis of submicron particles in heterogeneous biological fluids",
+  journal = "Journal of Controlled Release",
+  volume  = 220,
+  number  = "PtA",
+  year    = 2015
+}
 
-@article{zenodo,
-         title = {diff_classifier},
-         author = {Curtis C, Rokem A, and Nance E},
-         publisher = {Zenodo},
-         year = {2018},
-         doi= {10.5281/zenodo.1413740}
-         }
+@MISC{zenodo,
+  title     = "diff_classifier",
+  author    = "Curtis C, Rokem A, and Nance E",
+  publisher = "Zenodo",
+  year      = 2018,
+  doi       = "10.5281/zenodo.1413740"
+}
