+++ paper.bib
@@ -34,7 +34,7 @@
 }
 
 @MISC{zenodo,
-  title     = "diff_classifier",
+  title     = "diff\_classifier",
   author    = "Curtis C, Rokem A, and Nance E",
   publisher = "Zenodo",
   year      = 2018,
