### 2018-09-12 b3113c7
Hunted down all possible math symbols and added 20624
