### 2018-09-12 #939 @arfon
@ccurtis7 - I'm trying to debug this myself. The issue seems to be with the YAML header. I'm wondering if there are some hidden characters in there or something...
