### 2018-09-12 #939 @arfon
@ccurtis7 - the issue is with the underscore in the title. Please change:

```
title: 'diff_classifier: Parallelization of multi-particle tracking videos'

to

title: 'diff\_classifier: Parallelization of multi-particle tracking videos'
```
