### 2018-09-12 #939 @arfon
Also, I should try and sanitize this kind of input as it's passed to Pandoc. This isn't the first time it's caught us out...
