### 2018-09-12 13f6f9b
Fixed YAML header in paper.md
