+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'diff$\_$classifier: Parallelization of multi-particle tracking video analyses'
+title: 'diff\_classifier: Parallelization of multi-particle tracking video analyses'
 tags:
 - Python
 - multi-particle tracking
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-The [diff$\_$classifier](https://github.com/ccurtis7/diff_classifier) package seeks
+The [diff_classifier](https://github.com/ccurtis7/diff_classifier) package seeks
 to address the issue of scale-up in multi-particle tracking (MPT) analyses via a
 parallelization approach. MPT is a powerful analytical tool that has been used
 in fields ranging from aeronautics to oceanography to biomedical engineering
@@ -45,7 +45,7 @@ short fast-moving trajectories and could result in empty datasets [@Wang:2015].
 Due to variations in experimental conditions and image quality, user-selected
 tracking parameters can vary widely from video to video. As parameter selection
 can also vary from user to user, this also brings up the issue of
-reproducibility. diff$\_$classifier addresses these issues with regression
+reproducibility. diff_classifier addresses these issues with regression
 tools to predict input tracking parameters and parallelized script-based
 implementations in Amazon Web Services (AWS) S3 and Batch services. By manually
 tracking a small subset of the entire video dataset to be analyzed (5-10 videos
@@ -53,7 +53,7 @@ per experiment), users can predict tracking parameters based on intensity
 distributions of input images. This can simultaneously reduce
 time-to-first-result in MPT workflows and provide reproducible MPT results.
 
-diff$\_$classifier also includes downstream MPT analysis tools including mean
+diff_classifier also includes downstream MPT analysis tools including mean
 squared displacement and feature calculations, visualization tools, and a
 principle component analysis implementation. MPT is commonly used to calculate
 and report ensemble-averaged diffusion coefficients of nanoparticles and other
@@ -63,14 +63,14 @@ analysis to individual particle trajectories. By including a variety of features
 resolution, users can implement a range of data science analysis techniques to
 their MPT datasets.
 
-The source code for diff$\_$classifier has been archived to Zenodo with the
+The source code for diff_classifier has been archived to Zenodo with the
 linked DOI: [@zenodo]
 
 
 # Acknowledgements
 
 The authors would like to thank the eScience Institute for the resources and
-expertise provided through the Incubator Program that made diff$\_$classifier
+expertise provided through the Incubator Program that made diff_classifier
 possible. The authors would also like to thank funding from the National
 Institute of General Medical Sciences 1R35 GM124677-01 (E. Nance).
 
