### 2018-09-12 c8f9e85
Made same changes in paper.bib
