+++ paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Pulford:2005,
   author = "Pulford G W",
   title = "Taxonomy of multiple target tracking methods",
-  journal = "Iee Proceedings$-$Radar Sonar and Navigation",
+  journal = "Iee Proceedings - Radar Sonar and Navigation",
   volume = "152",
   number = "5",
   year   = "2005"
@@ -18,7 +18,7 @@
 
 @ARTICLE{Tineves:2017,
   author  = "Tineves J, Perry N, Schindelin J, Hoopes G M, Reynolds G D, Laplantine E, Bednarek S Y, Shorte S L, and Eliceiri K W",
-  title   = "TrackMate$\colon$ An open and extensible platform for single$-$particle tracking",
+  title   = "TrackMate$\colon$ An open and extensible platform for single-particle tracking",
   journal = "Methods",
   volume  = "115",
   year    = "2017"
@@ -34,9 +34,9 @@
 }
 
 @MISC{zenodo,
-  title     = "diff$\_$classifier",
+  title     = "diff\_classifier",
   author    = "Curtis C, Rokem A, and Nance E",
   publisher = "Zenodo",
   year      = "2018",
-  doi       = "10.5281$/$zenodo$\ldotp$1413740"
+  doi       = "10.5281/zenodo.1413740"
 }
