### 2018-09-12 #939 @ccurtis7
I'm not sure why it's still not working after I changed the YAML file. I'm unfamiliar with Latek, but I tried checking a [complete list](https://www.sharelatex.com/learn/Errors/Missing_$_inserted) of symbols that need to be written as math. Is the colon throwing it as well?
