### 2018-09-12 #939 @arfon
@ccurtis7 - I'm wondering if the new command `@whedon generate pdf from branch...` is being weird. I can compile your changes on the `joss_paper` branch locally OK.

Perhaps merge what you have in the `joss_paper` branch into `master` and just try `@whedon generate pdf`?

Sorry for the both here. I've only recently added the ability for Whedon to compile papers from a different branch and it's quite possibly a little buggy...
