### 2018-09-12 19da4b3
A few small edits in the paper.
