+++ paper.bib
@@ -42,7 +42,7 @@
 }
 
 
-@InProceedings{ adam_richie-halford-proc-scipy-2018,
+@InProceedings{ cloudknot,
   author    = { {A}dam {R}ichie-{H}alford and {A}riel {R}okem },
   title     = { {C}loudknot: {A} {P}ython {L}ibrary to {R}un your {E}xisting {C}ode on {A}{W}{S} {B}atch },
   booktitle = { {P}roceedings of the 17th {P}ython in {S}cience {C}onference },
