### 2018-09-12 #939 @ccurtis7
A few minor edits from @arokem
