### 2018-09-14 a49253f
New temp data storage folders
