+++ diff_classifier/data/trackmate_template2.py
@@ -1,157 +1,0 @@
-import sys
-from ij import IJ, ImagePlus, ImageStack
-import fiji.plugin.trackmate.Settings as Settings
-import fiji.plugin.trackmate.Model as Model
-import fiji.plugin.trackmate.SelectionModel as SelectionModel
-import fiji.plugin.trackmate.TrackMate as TrackMate
-import fiji.plugin.trackmate.Logger as Logger
-import fiji.plugin.trackmate.detection.DetectorKeys as DetectorKeys
-import fiji.plugin.trackmate.detection.DogDetectorFactory as DogDetectorFactory
-import fiji.plugin.trackmate.tracking.sparselap.SparseLAPTrackerFactory as SparseLAPTrackerFactory
-import fiji.plugin.trackmate.tracking.LAPUtils as LAPUtils
-import fiji.plugin.trackmate.visualization.hyperstack.HyperStackDisplayer as HyperStackDisplayer
-import fiji.plugin.trackmate.features.FeatureFilter as FeatureFilter
-import fiji.plugin.trackmate.features.FeatureAnalyzer as FeatureAnalyzer
-import fiji.plugin.trackmate.features.spot.SpotContrastAndSNRAnalyzerFactory as SpotContrastAndSNRAnalyzerFactory
-import fiji.plugin.trackmate.action.ExportStatsToIJAction as ExportStatsToIJAction
-import fiji.plugin.trackmate.io.TmXmlReader as TmXmlReader
-import fiji.plugin.trackmate.action.ExportTracksToXML as ExportTracksToXML
-import fiji.plugin.trackmate.io.TmXmlWriter as TmXmlWriter
-import fiji.plugin.trackmate.features.ModelFeatureUpdater as ModelFeatureUpdater
-import fiji.plugin.trackmate.features.SpotFeatureCalculator as SpotFeatureCalculator
-import fiji.plugin.trackmate.features.spot.SpotContrastAndSNRAnalyzer as SpotContrastAndSNRAnalyzer
-import fiji.plugin.trackmate.features.spot.SpotIntensityAnalyzerFactory as SpotIntensityAnalyzerFactory
-import fiji.plugin.trackmate.features.track.TrackSpeedStatisticsAnalyzer as TrackSpeedStatisticsAnalyzer
-import fiji.plugin.trackmate.features.track.TrackDurationAnalyzer as TrackDurationAnalyzer
-import fiji.plugin.trackmate.util.TMUtils as TMUtils
-
-
-# Get currently selected image
-#imp = WindowManager.getCurrentImage()
-#imp = IJ.openImage('{target_file}')
-#IJ.run(imp, "Properties...", "channels=1 slices=1 frames=651 unit=pixel pixel_width=1.0000 pixel_height=1.0000 voxel_depth=1.0000");
-#imp = IJ.openImage('/home/ubuntu/data/RED_nPEG_37C_pH72_S1_1_1_2.tif')
-imp = IJ.openImage('http://fiji.sc/samples/FakeTracks.tif')
-#imp.show()
-
-
-#-------------------------
-# Instantiate model object
-#-------------------------
-
-model = Model()
-
-# Set logger
-model.setLogger(Logger.IJ_LOGGER)
-
-#------------------------
-# Prepare settings object
-#------------------------
-
-settings = Settings()
-settings.setFrom(imp)
-
-# Configure detector
-settings.detectorFactory = DogDetectorFactory()
-settings.detectorSettings = {{
-    DetectorKeys.KEY_DO_SUBPIXEL_LOCALIZATION : True,
-    DetectorKeys.KEY_RADIUS : {radius},
-    DetectorKeys.KEY_TARGET_CHANNEL : 1,
-    DetectorKeys.KEY_THRESHOLD : {threshold},
[…]
