### 2018-09-20 5df3438
Small edits to functions.
