### 2018-09-22 #939 @ccurtis7
Others I think that might be good fits, @stsievert and @vsoch
