### 2018-09-25 2344a00
Added mean features
