### 2018-09-25 f8a98a6
Fixed test functions.
