### 2018-09-25 765e910
Fixed tests
