### 2018-09-30 #989 @whedon
[REVIEW]: Diff_classifier: Parallelization of multi-particle tracking video analyses
**Submitting author:** @ccurtis7 (<a href="http://orcid.org/0000-0001-6312-392X">Chad Curtis</a>)
**Repository:** <a href="https://github.com/ccurtis7/diff_classifier" target ="_blank">https://github.com/ccurtis7/diff_classifier</a>
**Version:** v0.1-beta

- [x] **Version:** v0.1-beta

    * see https://github.com/ccurtis7/diff_classifier/pull/18
