### 2018-10-02 296128e
Small fixes to pca analysis.
