### 2018-10-11 8f6e74d
Updating my branch of diff_classifier
