### 2018-10-25 afc31ec
New experiments and BF analysis.
