### 2018-11-13 49324dc
Add pip install note
