### 2018-11-13 a336d3d
Update getting_started.rst
