+++ doc/source/getting_started.rst
@@ -26,6 +26,12 @@ Running the setup file will install needed dependencies, including Fiji:
 
   python3 setup.py develop
 
+This allows you to test the installation:
+
+.. code-block:: python
+
+  cd diff_classifier/diff_classifier/tests
+  py.test
 
 You can install diff_classifier from the `Github repository
 <https://github.com/ccurtis7/diff_classifier>`_.  This will install
