### 2018-11-13 28a4091
Add Shapely to reqiurements
