### 2018-11-13 fde7086
MAINT: get requirements working
