+++ requirements.txt
@@ -1,4 +1,3 @@
-pip
 coverage==4.5.1
 coveralls
 pytest==3.5.1
