### 2018-11-25 #989 @stsievert
@ccurtis7 could you add DOIs for first 4 papers in `papers.bib`? i.e., the `Chenouard:2014` reference should have a DOI of https://doi.org/10.1038/nmeth.2808.
