+++ diff_classifier/imagej.py
@@ -99,7 +99,7 @@ def mean_intensity(local_im, frame=0):
 
 
 def track(target, out_file, template=None, fiji_bin=None,
-          tparams={'radius': 3.0, 'threshold': 0.0, 'do_median_filtering': False,
+          tparams={'frames': 651, 'radius': 3.0, 'threshold': 0.0, 'do_median_filtering': False,
            'quality': 15.0, 'xdims': (0, 511), 'ydims': (1, 511),
            'median_intensity': 300.0, 'snr': 0.0, 'linking_max_distance': 6.0,
            'gap_closing_max_distance': 10.0, 'max_frame_gap': 3,
@@ -151,6 +151,16 @@ def track(target, out_file, template=None, fiji_bin=None,
         Lower duration cutoff in frames for trajectory filtering.
 
     """
+    
+    tdefault = {'frames': 651, 'radius': 3.0, 'threshold': 0.0,
+                'do_median_filtering': False, 'quality': 15.0, 'xdims': (0, 511),
+                'ydims': (1, 511), 'median_intensity': 300.0, 'snr': 0.0,
+                'linking_max_distance': 6.0, 'gap_closing_max_distance': 10.0,
+                'max_frame_gap': 3, 'track_duration': 20.0}
+    for key in tdefault:
+        if key not in tparams.keys():
+            tparams[key] = tdefault[key]
+    
     if template is None:
         template = op.join(op.split(dc.__file__)[0],
                            'data',
@@ -166,7 +176,7 @@ def track(target, out_file, template=None, fiji_bin=None,
     script = ''.join(open(template).readlines())
     tpfile = tempfile.NamedTemporaryFile(suffix=".py")
     fid = open(tpfile.name, 'w')
-    fid.write(script.format(target_file=target, radius=str(tparams['radius']),
+    fid.write(script.format(target_file=target, frames=str(tparams['frames']), radius=str(tparams['radius']),
                             threshold=str(tparams['threshold']),
                             do_median_filtering=str(tparams['do_median_filtering']),
                             quality=str(tparams['quality']),
