+++ diff_classifier/data/trackmate_template3.py
@@ -29,8 +29,7 @@ import fiji.plugin.trackmate.util.TMUtils as TMUtils
 # Get currently selected image
 #imp = WindowManager.getCurrentImage()
 imp = IJ.openImage('{target_file}')
-IJ.run(imp, "Properties...", "channels=1 slices=1 frames=651 unit=pixel pixel_width=1.0000 pixel_height=1.0000 voxel_depth=1.0000");
-#imp = IJ.openImage('/home/ubuntu/data/RED_nPEG_37C_pH72_S1_1_1_2.tif')
+IJ.run(imp, "Properties...", "channels=1 slices=1 frames={frames} unit=pixel pixel_width=1.0000 pixel_height=1.0000 voxel_depth=1.0000");
 #imp = IJ.openImage('http://fiji.sc/samples/FakeTracks.tif')
 #imp.show()
 
