### 2018-11-26 2bf745e
Added dios to paper.bib
