+++ paper.bib
@@ -4,7 +4,8 @@
   journal = "Iee Proceedings - Radar Sonar and Navigation",
   volume = "152",
   number = "5",
-  year   = "2005"
+  year   = "2005",
+  doi    = "10.1049/ip-rsn:20045064"
 }
 
 @ARTICLE{Chenouard:2014,
@@ -13,7 +14,8 @@
   journal = "Nature Methods",
   volume  = "11",
   number  = "3",
-  year    = "2014"
+  year    = "2014",
+  doi     = "10.1038/nmeth.2808"
 }
 
 @ARTICLE{Tineves:2017,
@@ -21,7 +23,8 @@
   title   = "TrackMate$\colon$ An open and extensible platform for single-particle tracking",
   journal = "Methods",
   volume  = "115",
-  year    = "2017"
+  year    = "2017",
+  doi     = "10.1016/j.ymeth.2016.09.016"
 }
 
 @ARTICLE{Wang:2015,
@@ -30,7 +33,8 @@
   journal = "Journal of Controlled Release",
   volume  = "220",
   number  = "PtA",
-  year    = "2015"
+  year    = "2015",
+  doi     = "10.1016/j.jconrel.2015.10.021"
 }
 
 @MISC{zenodo,
