### 2018-12-03 ccurtis7/diff_classifier#22 @stsievert
No Fiji installation does not raise an error
I am testing for if Fiji not installed. When I run `diff_classifier_demo.ipynb`, I run

``` python
>>> # Cell 5
>>> ij.track(filename1, outfile, template=None, fiji_bin=None, tparams=tparams)
... print("Done with tracking.  Should output file of name {}".format(outfile))
/Applications/Fiji.app/Contents/MacOS/ImageJ-macosx --ij2 --headless --run /var/folders/_p/t_mfcz_90vx3shw664l6kpbr0000gq/T/tmp1fr_3dgs.py
Done with tracking.  Should output file of name Traj_FakeTracks.csv

>>> # part of Cell 6
>>> traj = ut.csv_to_pd(outfile)
No data in csv file
```

There is no error raised here. The notebook eventually raises a `ValueError: max() arg is an empty sequence` because no data is in the dataframe.
