### 2018-12-03 ccurtis7/diff_classifier#22 @stsievert
I think this is fixed by having a hard dependency on `fijibin`. It installs quickly, and after running `pip install fijibin`, I can do

``` python
>>> import fijibin
Fiji missing in /Users/scott/.bin/Fiji-20141125.app
Downloaded 100 percent
Extracting zip
>>> fijibin.FIJI_ROOT
'/Users/scott/.bin/Fiji-20141125.app'
```

This provides a path to an executable, as mentioned on https://ccurtis7.github.io/diff_classifier/getting_started.html. I think this would resolve the issue.
