### 2018-12-03 ccurtis7/diff_classifier#22 @ccurtis7
So, to fix this I need to:

(1) include `fijibin` in the `setup.py` file and
(2) edit the `track` function to use `fijibin.FIJI_ROOT` to point to the binary, instead of the current semi-manual `fiji_bin` variable.

Would that be enough?
