### 2018-12-03 ccurtis7/diff_classifier#22 @stsievert
> 1. include fijibin in the setup.py file and
> 2. edit the track function to use fijibin.FIJI_ROOT to point to the binary, instead of the current semi-manual fiji_bin variable.

Yes, that's what I would do. It might be `fijibin.BIN` instead of `fijibin.FIJI_ROOT`.

I'd probably write the code so Fiji is only downloaded when needed and to allow users to use their own Fiji.
