### 2018-12-03 3cc2e97
Modified download behavior of fiji
