### 2018-12-03 ccurtis7/diff_classifier#22 @ccurtis7
OK, so I used `fijibin.BIN` like you suggested, and made edits to `setup.py` and `track`. I still use the `fiji_bin` object to reference the binary.

I think it would be best if `fijibin` were able to detect a pre-existing download of fiji, but I can't find any reference to such functionality. As-is, if `fiji_bin` is set to `None`, fiji will be downloaded. If the user wants to use a pre-existing download of fiji, they need to manually reference define the binary in `fiji_bin`.
