### 2019-01-07 #989 @ccurtis7
I forgot to mention that I went and updated the 'papers.bib' file as requested.
