### 2019-01-07 #989 @ccurtis7
@stsievert I hope you had a good winter break! I noticed that you have checked everything except installation and performance. I believe I have addressed all the issues you brought up.

Let me know if I can do anything else to finish up!
