### 2019-01-09 #989 @stsievert
Thanks for the ping @ccurtis7, and my apologies for the excessive delay.

I think this paper is coming to an acceptable state. Past the issues/PRs I've filed, I think the only blocker is verification of the CloudKnot parallelization claims. I'll try to do that this weekend.
