### 2019-01-09 5617c04
Updated default behavior of imagej.track.
