### 2019-01-09 7abdd1f
Updated py.test version to get travis to pass.
