### 2019-01-09 d494e0a
Updated py.test version to get travis to pass.
