+++ requirements.txt
@@ -1,4 +1,4 @@
-coverage==4.5.1
+coverage==4.5.2
 coveralls
 pytest==4.1.0
 numpy
