### 2019-01-09 b5c0f60
Modified default fiji_bin behavior in track.imagej.
