### 2019-01-09 30ac974
Modified setup.py to download Fiji.
