+++ setup.py
@@ -1,4 +1,5 @@
 from setuptools import setup, find_packages
+import setuptools.command.build_py
 import os
 # try:
 #    import fijibin
@@ -28,10 +29,20 @@ opts = dict(name=NAME,
             install_requires=REQUIRES,
             package_data=PACKAGE_DATA)
 
+class FijiCommand(setuptools.command.build_py.build_py):
+    """Downloads and unzips current version of Fiji."""
+    
+    def run(self):
+        self.run_command('cd ~')
+        self.run_command('wget https://downloads.imagej.net/fiji/latest/fiji-linux64.zip')
+        self.run_command('unzip fiji-linux64.zip')
+        setuptools.command.build_py.build_py.run(self)
+
 
 if __name__ == '__main__':
     setup(**opts)
     try:
-        import fijibin
+        #import fijibin
+        setup(cmdclass={'build_py': FijiCommand})
     except:
         print('Import error. Install Fiji manually.')
