+++ _sources/_sources/api/index.rst.txt
@@ -1,0 +1,15 @@
+API
+===
+
+This is the diff_classifier documentation, auto generated from the source code.
+
+.. toctree::
+
+    utils
+    msd
+    imagej
+    features
+    aws
+    heatmaps
+    pca
+    knotlets
