+++ _sources/_sources/cloudknot_parallelization.rst.txt
@@ -1,0 +1,13 @@
+.. _cloudknot-parallelization-label:
+
+Cloudknot parallelization
+=========================
+
+Diff_classifier also includes example implementations of parallelizing tracking
+analysis and visualization with Cloudknot, a package developed by Adam
+Richie-Halford.  Checkout out the documentation
+`here <https://richford.github.io/cloudknot/>`_. Some demonstration functions
+are included in the knotlets module, but can easily be built upon for custom
+parallelization particle tracking workflows. See
+`this notebook <https://github.com/ccurtis7/diff_classifier/blob/master/notebooks/cloudknot_parallelization_demo.ipynb>`_
+for an example implementation.
