+++ _sources/_sources/api/pca.rst.txt
@@ -1,0 +1,9 @@
+:mod:`diff_classifier.pca`
+=================================
+
+:mod:`diff_classifier.pca`
+---------------------------------
+.. automodule:: diff_classifier.pca
+    :members:
+    :undoc-members:
+    :show-inheritance:
