### 2019-01-13 f433b48
MAINT: clean getting Fiji binary
