+++ setup.py
@@ -29,15 +29,6 @@ opts = dict(name=NAME,
             install_requires=REQUIRES,
             package_data=PACKAGE_DATA)
 
-class FijiCommand(setuptools.command.build_py.build_py):
-    """Downloads and unzips current version of Fiji."""
-    
-    def run(self):
-        self.run_command('cd ~')
-        self.run_command('wget https://downloads.imagej.net/fiji/latest/fiji-linux64.zip')
-        self.run_command('unzip fiji-linux64.zip')
-        setuptools.command.build_py.build_py.run(self)
-
 
 if __name__ == '__main__':
     setup(**opts)
