### 2019-01-21 2377543
updating notebooks and adding PNN binning functions
