+++ diff_classifier/PNNs.py
@@ -1,0 +1,160 @@
+import os
+import os.path as op
+import boto3
+import numpy as np
+import pandas as pd
+import matplotlib
+import matplotlib.pyplot as plt
+import skimage.io
+import skimage.filters
+from skimage.filters import try_all_threshold
+from diff_classifier import aws
+import matplotlib.image as mpimg
+from scipy import ndimage as ndi
+from skimage.transform import resize
+import numpy.ma as ma
+from skimage import morphology
+from scipy.ndimage.morphology import distance_transform_edt as Euclidean
+
+def PNN_binning(raw_img, input_df, num_bins=7, threshold='Otsu', min_obj_size=3000, bin_cuts=[0,3.57,7.14,14.28,142.8,714.3], \
+                feat='Deff1', box_plot=True):
+    
+    
+    # First, you have to scale the PNN confocal image so it aligns with the size of the MPT video
+    # (usually, it must be scaled from a 512x512 image to a 2048x2048 image)
+    img_scaled = resize(raw_img[0,:,:],(2048,2048))
+    scaled_plot = plt.imshow(img_scaled, cmap='gray')
+    plt.show()
+    
+    # this thresholds the scaled image using the user-input threshold type, converting the image to a binary
+    
+    if threshold == 'Otsu':
+        thresh = skimage.filters.threshold_otsu(img_scaled)
+    elif threshold == 'Mean':
+        thresh = skimage.filters.threshold_mean(img_scaled)
+    elif threshold == 'Triangle':
+        thresh = skimage.filters.threshold_triangle(img_scaled)
+    elif threshold == 'Yen':
+        thresh = skimage.filters.threshold_yen(img_scaled)
+    elif threshold == 'Li':
+        thresh = skimage.filters.threshold_li(img_scaled)
+    elif threshold == 'Isodata':
+        thresh = skimage.filters.threshold_isodata(img_scaled)
+    else:
+        thresh = skimage.filters.threshold_minimum(img_scaled)
+    
+    pnnbinary = img_scaled > thresh
+    
+    # these next few lines of code fill any holes existing in the binary image and removes any 
+    # objects smaller than the user-defined minimum object size
+    
+    pnnbinary_filled = ndi.binary_fill_holes(pnnbinary)
+    binary_filledplot = plt.imshow(pnnbinary_filled, cmap='gray')
+    plt.show()
+    pnn_clean = morphology.remove_small_objects(pnnbinary_filled, min_size=min_obj_size)
+    pnn_clean_plot = plt.imshow(pnn_clean, cmap='gray')
+    plt.show()
+    
+    # Now, the Euclidean distance (to the nearest PNN) is caluclated at each pixel location in the 2048x2048 image
+    
+    euc_img = Euclidean(1-pnn_clean) #1- represents going outwards from the cells
+    plt.imshow(euc_img)
+    plt.show
+    
+    # With the Euclidean distance matrix generated, we now switch over to dataframe modification
+    
+    raw_df = input_df
+    raw_df['Euc'] = np.nan
+    raw_df['Bin'] = np.nan
+    
+    tot_traj = int(max(raw_df['Track_ID']))
+    
+    # The following for loop goes through all trajectories in the features dataframe, calculates the Euclidean 
+    # distance from the centroid of the trajectory, and bins the trajectories based on those distances
+    
+    # don't want to deal with it now, but will have to adjust this section of code to be able to adjust the 
+    # number of elif statements to match the number of bins chosen by the user
+    
[…]
