### 2019-01-24 9fd93a6
Modified random walker to include mask.
