+++ diff_classifier/msd.py
@@ -16,6 +16,7 @@ from scipy import interpolate
 import matplotlib.pyplot as plt
 from matplotlib.pyplot import cm
 import diff_classifier.aws as aws
+from scipy.ndimage.morphology import distance_transform_edt as eudist
 
 
 def nth_diff(dataframe, n=1, axis=0):
@@ -717,7 +718,7 @@ def plot_all_experiments(experiments, bucket='ccurtis.data', folder='test',
     """
 
     n = len(experiments)
-    
+
     if labels==None:
         labels = experiments
 
@@ -750,23 +751,23 @@ def plot_all_experiments(experiments, bucket='ccurtis.data', folder='test',
 
         if exponential:
             ax.plot(xpos, np.exp(geo[counter]), c=c, linewidth=6,
-                       label=labels[counter])
+                    label=labels[counter])
             ax.fill_between(xpos, np.exp(geo[counter] - 1.96*gstder[counter]),
                             np.exp(geo[counter] + 1.96*gstder[counter]),
                             color=c, alpha=0.4)
 
         else:
             ax.plot(xpos, geo[counter], c=c, linewidth=6,
-                       label=labels[counter])
-            ax.fill_between(xpos, geo[counter] - 1.96*gstder[counter], 
-                       geo[counter] + 1.96*gstder[counter], color=c,
-                       alpha=0.4)
+                    label=labels[counter])
+            ax.fill_between(xpos, geo[counter] - 1.96*gstder[counter],
+                            geo[counter] + 1.96*gstder[counter], color=c,
+                            alpha=0.4)
 
         counter = counter + 1
-    
+
     if log:
-        ax.set_xscale( "log" )
-        ax.set_yscale( "log" )
+        ax.set_xscale("log")
+        ax.set_yscale("log")
 
     plt.legend(frameon=False, loc=2, prop={'size': 16})
 
@@ -775,7 +776,62 @@ def plot_all_experiments(experiments, bucket='ccurtis.data', folder='test',
         aws.upload_s3(outfile, folder+'/'+outfile, bucket_name=bucket)
 
 
-def random_walk(nsteps=100, seed=1, start=(0, 0)):
+def checkerboard_mask(dims=(512, 512), squares=50, width=25):
+    """Creates a 2D Boolean checkerboard mask
+
+    Creates a Boolean array of evenly spaced squares.
+    Whitespace is set to True.
+
+    Parameters
+    ----------
+
+    dims : tuple of int
+        Dimensions of desired Boolean array
+    squares : int
+        Dimensions of in individual square in array
+    width : int
+        Dimension of spacing between squares
+
+    Returns
+    ----------
+
+    zeros : numpy.ndarray of bool
+        2D Boolean array of evenly spaced squares
+
+    """
+    zeros = np.zeros(dims) == 0
+    square_d = squares
+
+    loy = width
+    hiy = loy + square_d
+
+    for j in range(50):
+
+        lox = width
+        hix = lox + square_d
+        for i in range(50):
+
+            if hix < 512 and hiy < 512:
+                zeros[loy:hiy, lox:hix] = False
+            elif hix < 512:
+                zeros[loy:512-1, lox:hix] = False
+            elif hiy < 512:
+                zeros[loy:hiy, lox:512-1] = False
+            else:
+                zeros[loy:512-1, lox:512-1] = False
+                break
+
+            lox = hix + width
+            hix = lox + square_d
+
+        loy = hiy + width
+        hiy = loy + square_d
+
+    return zeros
+
+
+def random_walk(nsteps=100, seed=None, start=(0, 0), step=1, mask=None,
+                stuckprob=0.5):
     """Creates 2d random walk trajectory.
 
[…]
