### 2019-01-24 ccurtis7/diff_classifier#28 @ccurtis7
Checking Cloudknot functionality
Hey @stsievert, any chance you can take a look at this this weekend?
