### 2019-01-24 ccurtis7/diff_classifier#28 @stsievert
I put it on today's checklist this morning – I'd like to get this done today. Thanks for the followup.
