### 2019-01-24 ccurtis7/diff_classifier#28 @stsievert
I ran into some issues with the cloudknot parallelization demo. The issues I ran into:

* copying the S3 files to my bucket took about 15 minutes (20GB at about 20 MiB/s)
* I had issues with the upload/download of the TIFFs in `diff_classifier.cloudknot.split` that took a long time to resolve (crappy wifi, moving to cluster, using cluster disk space optimally).

It'd be nice if this could be reduced (it took me over an hour to resolve), but it's not blocking on the JOSS review.

---

Also, small nit: it'd be nice if the `bucket` argument of `diff_classifier.knotlets.split` be required. Currently, it defaults to `nancelab.publicfiles` bucket, which I don't have access to write to.
