### 2019-01-26 ccurtis7/diff_classifier#28 @stsievert
It's nice to have a real-world analysis use case in the Cloudknot demo. I didn't express that in https://github.com/ccurtis7/diff_classifier/issues/28#issuecomment-457394328. I think this demo should remain as is, and it'd be nice to add one that's a little quicker to run.
