### 2019-01-26 ccurtis7/diff_classifier#28 @ccurtis7
Do you have any recommendations about speeding up both the copying across buckets and downloading from AWS? In the experimental workflow, uploading to AWS is always the bottleneck, but the downloads and transfer seemed fine as long as you were working from an EC2 instance. Overall though, doing the entire analysis in AWS is lightyears faster than what I could do locally.

---

To make the `bucket` argument required, do I just remove the default option?

---

Oh, interesting, were you able to get the entire notebook to run in Python3? I thought it had to do something with the `subprocess` function having different behavior, but I'm not quite sure. @arokem, do you remember why we had to insist on running the Cloudknot submission in Python2?

As for the buggy behavior of AWS, I know how to navigate around that problem. First, you have to double check that you are in the right region. But even if you are in the right region, it still spend you to the "first time" page sometimes. If you click on "Get started" and then "Skip Wizard", it will take you to the correct interface.
