### 2019-01-27 ccurtis7/diff_classifier#28 @arokem
Regarding Python 2/3: cloudknot definitely works on both. If I remember correctly, the reason we wanted to work in Python 2 was that we were working on AWS machines that had Python 2.7 installed on them per default and we needed to work under the same environment that root was using to be able to use docker. So I think that this is not a fundamental limitation, but rather a limitation to do with the way we were running this on AWS machines.
