### 2019-01-28 ccurtis7/diff_classifier#28 @stsievert
> this is not a fundamental limitation, but rather a limitation to do with the way we were running this on AWS machines.

It'd be nice to resolve this. What would be required to change it?

> First, you have to double check that you are in the right region.

Thanks @ccurtis7! I forgot about that annoying limitation. I'll hopefully give this another go soon.
