### 2019-01-30 ccurtis7/diff_classifier#28 @arokem
On Mon, Jan 28, 2019 at 9:17 AM Scott Sievert <notifications@github.com>
wrote:

> this is not a fundamental limitation, but rather a limitation to do with
> the way we were running this on AWS machines.
>
> It'd be nice to resolve this. What would be required to change it?
>
This is not a problem at all if you are running this on a non-linux system,
where docker doesn't require sudo. For example, I run Cloudknot on my mac
on both Python 2 and Python 3, with different virtualenvs. Alternatively,
you can run your Python process (e.g., Jupyter) as sudo on linux. Either
way, you need to have the permissions to run docker.

Does that make sense?

> First, you have to double check that you are in the right region.
>
> Thanks @ccurtis7 <https://github.com/ccurtis7>! I forgot about that
> annoying limitation. I'll hopefully give this another go soon.
>
