### 2019-01-30 ccurtis7/diff_classifier#28 @stsievert
Why does choice of operating system/sudo privileges enter the picture at all? Is it because sudo will use the `python` command on *nix, which defaults to Python 2.7 on some distros ([PEP 394])? Would this still be an issue if `python` ran Python 3.x?

[PEP 394]:https://www.python.org/dev/peps/pep-0394/
