### 2019-01-30 ccurtis7/diff_classifier#28 @arokem
On Wed, Jan 30, 2019 at 9:57 AM Scott Sievert <notifications@github.com>
wrote:

> Why does choice of operating system/sudo privileges enter the picture at
> all? Is it because sudo will use the python command on *nix, which
> defaults to Python 2.7 on some distros (PEP 394
> <https://www.python.org/dev/peps/pep-0394/>)? Would this still be an
> issue if python ran Python 3.x?
>
>
Because Cloudknot uses Docker under the hood (and that's an AWS Batch
requirement) and Docker requires sudo priviliges on linux, but not on other
OS (e.g., on the Mac).
