### 2019-01-31 ccurtis7/diff_classifier#28 @ccurtis7
I went through the notebook and tried running it in Python2 and Python3, and it errors out in Python3 when I run `cloudknot.DockerImage`. It appears that in Python3, there no longer is a `base_image` variable. @arokem and I built a Docker Image that has Fiji properly installed prior to running the code. Interesting that the Python2 and Python3 versions of `cloudknot.DockerImage` have different input variables.

As we have cleaned up the Fiji installation process when using `diff_classifier` perhaps we don't need to use a base image anymore? I'm going to try testing it out.

```
