### 2019-01-31 4e05daa
Fixed _get_fiji installation bug.
