### 2019-01-31 f4fc8c5
Fixed another _get_fiji bug
