### 2019-02-01 a923ff7
Fixed _get_fiji to return proper path.
