### 2019-02-01 69f1ac9
Added another option to imagej.track to get around Trackmate bug.
