### 2019-02-01 ccurtis7/diff_classifier#28 @ccurtis7
Success! I was able to remove the `Python2 requirement` completely after a few modifications. I had to change the call to `cloudknot.DockerImage` in the demo notebook as explained above due to changes in behavior. There was also a new bug in `Trackmate` that made it so the difference of gaussians spot detector was erroring out, resulting in empty CSVs. Weird. I added additional functionality into `imagej.track` that allows the user to choose between difference of gaussians and log of gaussians so you can bypass this error.

I just need to update the documentation to remove the Python2 requirement, and we're all set. Let me know if you can get the demo notebook running to @stsievert
