### 2019-02-01 67a19b7
Updated documentation to remove Python2 req.
