+++ _sources/api/msd.rst.txt
@@ -1,0 +1,9 @@
+:mod:`diff_classifier.msd`
+=================================
+
+:mod:`diff_classifier.msd`
+---------------------------------
+.. automodule:: diff_classifier.msd
+    :members:
+    :undoc-members:
+    :show-inheritance:
