### 2019-02-06 66e4fb0
Fixed imagej Mac bug.
