### 2019-02-13 #989 @stsievert
Thanks for checking in. I’ve found a couple issues in CloudKnot, specifically on some dependencies (see https://github.com/ccurtis7/diff_classifier/issues/28 for some more detail). @ccurtis7 has resolved that issue, and I need to check the notebook again.

I don’t think I would have accepted the version of diff_classifier as submitted. I think now it’s closer to being ready.
