### 2019-02-13 #989 @ccurtis7
Hey @cMadan, @stsievert has really given some great help on improving the quality of the package, and I've tried to be as prompt as possible in addressing issues. As he mentioned, I think there's only one last double check of the parallelization demo notebook, and it should be good to go.
