### 2019-02-25 ccurtis7/diff_classifier#28 @ccurtis7
Hey @stsievert, have you been able to check out the updated Cloudknot notebook yet?
