### 2019-02-26 ccurtis7/diff_classifier#30 @stsievert
No Fiji means infinite download loop in diff_classifier_demo.ipynb
When executing the cell

``` python
ij.track(filename1, outfile, template=None, fiji_bin=None, tparams=tparams)
print("Done with tracking.  Should output file of name {}".format(outfile))
```

it goes into an infinite download loop. Here's some output when that cell is run from the shell where Jupyter is running:

```
fiji-macosx.zip                               100%[==============================================================================================>] 318.48M   876KB/s    in 7m 30s

2019-02-25 19:52:29 (725 KB/s) - 'fiji-macosx.zip' saved [333947759/333947759]

Archive:  fiji-macosx.zip
  inflating: Fiji.app/db.xml.gz
  inflating: Fiji.app/Contents/Info.plist
  ...
  inflating: Fiji.app/java/macosx/adoptopenjdk-8.jdk/jre/Contents/MacOS/libjli.dylib
--2019-02-25 19:52:35--  https://downloads.imagej.net/fiji/latest/fiji-macosx.zip
Resolving downloads.imagej.net... 144.92.48.177
Connecting to downloads.imagej.net|144.92.48.177|:443... connected.
HTTP request sent, awaiting response... 200 OK
Length: 333947759 (318M) [application/zip]
Saving to: 'fiji-macosx.zip.1'
```

It looks like it unpacked Fiji.app successfully and placed it in `~/Fiji.app`, but it keeps on downloading.
