### 2019-02-26 ccurtis7/diff_classifier#30 @stsievert
It looks like the environment variable `FIJI_BIN` isn't set in https://github.com/ccurtis7/diff_classifier/blob/66e4fb01592b964b8b80bc09eb6edc0775f3da01/diff_classifier/imagej.py#L68
