### 2019-02-26 68a5331
Updated knotlets.py to have better defaults
