### 2019-02-26 41eef26
Modififed imagej._get_fiji to define FIJI_BIN variable
