### 2019-02-26 ccurtis7/diff_classifier#30 @ccurtis7
Ah, it's because I send the download to `~/Fiji.app`, but then search for it at `~/Applications/Fiji.app` on a Mac. Sorry about that-- I don't have access to a Mac to test run this. I've been using a Linux.

I did what you recommended and defined the `FIJI_BIN` environment variable after download. Now it should work.

https://github.com/ccurtis7/diff_classifier/blob/master/diff_classifier/imagej.py#L73-83
