### 2019-02-26 128aada
Modified heatmaps.py to limit RAM memory usage by plotting subset of data.
