### 2019-02-26 40152f3
Fixed sampling error in updates
