+++ diff_classifier/heatmaps.py
@@ -281,6 +281,10 @@ def plot_trajectories(prefix, resolution=512, rows=4, cols=4, upload=True,
     """
     merged = pd.read_csv('msd_{}.csv'.format(prefix))
     particles = int(max(merged['Track_ID']))
+    if particles > size:
+        size = particles - 1
+    else:
+        pass
     particles = np.linspace(0, particles, particles-1).astype(int)
     if subset:
         particles = np.random.choice(particles, size=size, replace=False)
@@ -472,6 +476,12 @@ def plot_individual_msds(prefix, x_range=100, y_range=20, umppx=0.16, fps=100.02
 
     fig = plt.figure(figsize=figsize)
     particles = int(max(merged['Track_ID']))
+
+    if particles > size:
+        size = particles - 1
+    else:
+        pass
+
     frames = int(max(merged['Frame']))
 
     y = merged['Y'].values.reshape((particles+1, frames+1))*umppx*umppx
