### 2019-02-26 ccurtis7/diff_classifier#30 @stsievert
Thanks – I'll test this later today to verify it works.
