### 2019-02-26 df3b5dd
Fixed heatmap test.
