### 2019-02-26 2bbc472
Fixed heatmap test.
