### 2019-02-27 ccurtis7/diff_classifier#28 @stsievert
I've ran through the notebook tonight, and produced the same plots as the notebook had originally.
