### 2019-02-27 ccurtis7/diff_classifier#28 @ccurtis7
Great! I'll close this one then.
