### 2019-02-27 4a9dea2
Addressed stsievert's issue about assumed values in knotlets.
