### 2019-02-27 9d093a2
Documentation clarifications
