### 2019-03-03 #989 @stsievert
I think diff_classifier is almost ready for acceptance. Here's a quick summary of the changes since the `v0.1-alpha` tag:

* Clean dependencies on Fiji, an image processing library (https://github.com/ccurtis7/diff_classifier/issues/22, https://github.com/ccurtis7/diff_classifier/issues/30)
* Clean confusion on Python 2/3 decadency on remote/local machines with Cloudknot (resolved in https://github.com/ccurtis7/diff_classifier/issues/28#issuecomment-459855740)
* Cleaning usage with `pip`: https://github.com/ccurtis7/diff_classifier/pull/18

Here are all the commits since the `v0.1-alpha` tag: https://github.com/ccurtis7/diff_classifier/compare/V0.1-alpha...master
