+++ _sources/api/index.txt
@@ -1,15 +1,0 @@
-API
-===
-
-This is the diff_classifier documentation, auto generated from the source code.
-
-.. toctree::
-
-    utils
-    msd
-    imagej
-    features
-    aws
-    heatmaps
-    pca
-    knotlets
