+++ _sources/_sources/getting_started.rst.txt
@@ -1,63 +1,0 @@
-.. _getting-started-label:
-
-Getting started with diff_classifier
-====================================
-
-Pip install
------------
-Currently, diff_classifier is not on PyPi, but the next version of
-diff_classifer will be uploaded to PyPi.
-
-Git clone
----------
-Users can clone a copy of diff_classifier with the command
-
-.. code-block:: python
-
-  git clone https://github.com/ccurtis7/diff_classifier.git
-
-Running the setup file will install needed dependencies, including Fiji:
-
-.. code-block:: python
-
-  python3 setup.py develop
-
-
-You can install diff_classifier from the `Github repository
-<https://github.com/ccurtis7/diff_classifier>`_.  This will install
-diff_classifier and its Python dependencies. Users will also need to have
-downloaded `Fiji <https://imagej.net/Fiji/Downloads>`_ in order to implement the
-Trackmate functionality.
-
-Checking Fiji installation
---------------------------
-The tracking portion of diff_classifier relies on the Fiji plugin `Trackmate
-<https://imagej.net/TrackMate>`_. The setup.py file should do this automatically.
-
-Otherwise, users can install Fiji manually at `here <https://fiji.sc/#download>`_.
-An example installation using the shell:
-
-.. code-block:: shell
-
-  wget https://downloads.imagej.net/fiji/latest/fiji-linux64.zip
-  unzip fiji-linux64.zip
-
-When using the diff_classifier.imagej module, default behavior looks for Fiji at
-~/Fiji.app/ImageJ-linux64.exe for Linux users or
-/Applications/Fiji.app/Contents/MacOS/ImageJ-macosx.exe for Apple users.
-Otherwise, users must manually specify the location using the fiji_bin variable.
-
-
-Note: Python2/Python3 compatibility
------------------------------------
-
-Use Python 3 for almost everything. Unfortunately, Python 2 must be used for
-one task: launching Cloudknot jobs. This means that Python 2 must be run
-locally when launching Cloudknot jobs, which launches an AWS Batch. This AWS
-Batch should use Python 3.
-
-Examples
---------
-You can find example implementations of diff_classifier in the `notebooks
-directory <https://github.com/ccurtis7/diff_classifier/tree/master/notebooks>`_
-on Github.
