### 2019-03-06 3001fa7
Added MPT examples to paper.md
