### 2019-03-06 e4e3057
Added JOSS badge to README
