### 2019-03-06 4895fe4
Adjusted figure height in README
