### 2019-03-06 292ca04
Adjusted figure height in README
