### 2019-03-12 692c806
Modified assemble_msds method of extracting rows and cols
