### 2019-03-12 29853a3
Modified assemble_msds so it doesn't error out with deep folder names
