### 2019-03-19 3a765f9
Added in info on adding new features to CONTRIBUTING.md
