### 2019-03-19 42719c0
Added link to scikit-learn on CONTRIBUTING.md for pull requests
