### 2019-03-31 #989 @stsievert
This review is complete! This software package is ready for acceptance. One question:

Significant changes have been made since the initial submission (some changes are in https://github.com/openjournals/joss-reviews/issues/989#issuecomment-469061459). I feel there should be a new release for this – I wouldn't have accepted `0.1-alpha`, but I'm willing to accept the `master` branch now. How should this be resolved?
