### 2019-04-03 #989 @cMadan
@stsievert, perfect, thank you for confirming! Thanks for highlighting this difference :). After I do a final pass of the submission, I will ask @ccurtis7 to mint a new release and archive the code (@ccurtis7, wait until I do a final pass though!) and this will then be the version that is accepted.

@ccurtis7, sit tight for now and I will let you know how the last steps proceed shortly.
