### 2019-04-07 #989 @cMadan
@ccurtis7, everything look good and I'm almost ready to accept. Can you mint a new release that includes the changes from the review process? I also need a DOI for an archived version of the code (i.e., from Zenodo or figshare).
