### 2019-04-07 #989 @ccurtis7
@cMadan, I have minted a new release, [v0.1-beta](https://github.com/ccurtis7/diff_classifier/releases/tag/v0.1-beta), and the DOI is 10.5281/zenodo.2631862
