### 2019-04-09 #989 @cMadan
@ccurtis7, it looks like a recent change broke the paper.md (or bib)
