### 2019-04-09 f616a1e
Attempted to fix broken paper.md file
