+++ paper.bib
@@ -45,13 +45,12 @@
   doi       = "10.5281/zenodo.1413740"
 }
 
-
-@InProceedings{ cloudknot,
-  author    = { {A}dam {R}ichie-{H}alford and {A}riel {R}okem },
-  title     = { {C}loudknot: {A} {P}ython {L}ibrary to {R}un your {E}xisting {C}ode on {A}{W}{S} {B}atch },
-  booktitle = { {P}roceedings of the 17th {P}ython in {S}cience {C}onference },
-  pages     = { 8 - 14 },
-  year      = { 2018 },
-  editor    = { {F}atih {A}kici and {D}avid {L}ippa and {D}illon {N}iederhut and {M} {P}acer },
-  doi       = { 10.25080/Majora-4af1f417-001 }
+@InProceedings{cloudknot,
+  author    = "Richie-Halford A, Rokem A",
+  title     = "Cloudknot$\colon$ A Python Library to Run your Existing Code on AWS Batch",
+  booktitle = "Proceedings of the 17th Python in Science Conference",
+  pages     = "8-14",
+  year      = "2018",
+  editor    = "Akici F, Lippa D, Niederhut D, Pacer M",
+  doi       = "10.25080/Majora-4af1f417-001"
 }
