### 2019-04-09 f41c58e
Eliminating paper.bib file as culprit.
