+++ paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 The [diff_classifier](https://github.com/ccurtis7/diff_classifier) package seeks
 to address the issue of scale-up in multi-particle tracking (MPT) analyses via a
 parallelization approach. MPT is a powerful analytical tool that has been used
-in fields ranging from aeronautics to oceanography [@Pulford:2005] allowing
+in fields ranging from aeronautics to oceanography allowing
 researchers to collect spatial and velocity information of moving objects from
 video datasets. Examples include:
 
@@ -39,16 +39,16 @@ video datasets. Examples include:
 * Assessing microrheological properties by tracking nanoparticle movement.
 
 While a variety of tracking algorithms are available to researchers
-[@Chenouard:2014], a common problem is that data analysis usually depends on the
+, a common problem is that data analysis usually depends on the
 use of graphical user interfaces, and relies on human input for accurate
 tracking. For example, particle detection often relies on the selection of a
 quality threshold, a numerical quantity distinguishing between “real” particles
-and “fake” particles [@Tineves:2017]. If this threshold is too high, false
+and “fake” particles . If this threshold is too high, false
 positive trajectories result in skewed MSD profiles, and in extreme cases, cause
 the code to crash due to a lack of convergence in the particle linking step. If
 the threshold is too low, trajectories will be cut short resulting in a bias
 towards short fast-moving trajectories and could result in empty datasets
-[@Wang:2015].
+.
 
 Due to variations in experimental conditions and image quality, user-selected
 tracking parameters can vary widely from video to video. As parameter selection
@@ -57,7 +57,7 @@ reproducibility. diff_classifier addresses these issues with regression tools to
 predict input tracking parameters and parallelized script-based implementations
 in Amazon Web Services (AWS), using the Simple Storage Service (S3) and Batch
 for data storage and computing, respectively, and relying on the Cloudknot
-software library for automating these interactions [@cloudknot]. By manually
+software library for automating these interactions . By manually
 tracking a small subset of the entire video dataset to be analyzed (5-10 videos
 per experiment), users can predict tracking parameters based on intensity
 distributions of input images. This can simultaneously reduce
@@ -74,7 +74,7 @@ resolution, users can implement a range of data science analysis techniques to
 their MPT datasets.
 
 The source code for diff_classifier has been archived to Zenodo with the
-linked DOI: [@zenodo].
+linked DOI: .
 
 
 # Acknowledgements
