+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'diff\_classifier: Parallelization of multi-particle tracking video analyses'
+title: 'diffclassifier: Parallelization of multi-particle tracking video analyses'
 tags:
 - Python
 - multi-particle tracking
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-The [diff_classifier](https://github.com/ccurtis7/diff_classifier) package seeks
+The [diffclassifier](https://github.com/ccurtis7/diff_classifier) package seeks
 to address the issue of scale-up in multi-particle tracking (MPT) analyses via a
 parallelization approach. MPT is a powerful analytical tool that has been used
 in fields ranging from aeronautics to oceanography allowing
@@ -53,7 +53,7 @@ towards short fast-moving trajectories and could result in empty datasets
 Due to variations in experimental conditions and image quality, user-selected
 tracking parameters can vary widely from video to video. As parameter selection
 can also vary from user to user, this also brings up the issue of
-reproducibility. diff_classifier addresses these issues with regression tools to
+reproducibility. diffclassifier addresses these issues with regression tools to
 predict input tracking parameters and parallelized script-based implementations
 in Amazon Web Services (AWS), using the Simple Storage Service (S3) and Batch
 for data storage and computing, respectively, and relying on the Cloudknot
@@ -63,7 +63,7 @@ per experiment), users can predict tracking parameters based on intensity
 distributions of input images. This can simultaneously reduce
 time-to-first-result in MPT workflows and provide reproducible MPT results.
 
-diff_classifier also includes downstream MPT analysis tools including mean
+diffclassifier also includes downstream MPT analysis tools including mean
 squared displacement and feature calculations, visualization tools, and a
 principle component analysis implementation. MPT is commonly used to calculate
 and report ensemble-averaged diffusion coefficients of nanoparticles and other
@@ -73,14 +73,14 @@ analysis to individual particle trajectories. By including a variety of features
 resolution, users can implement a range of data science analysis techniques to
 their MPT datasets.
 
-The source code for diff_classifier has been archived to Zenodo with the
+The source code for diffclassifier has been archived to Zenodo with the
 linked DOI: .
 
 
 # Acknowledgements
 
 The authors would like to thank the eScience Institute for the resources and
-expertise provided through the Incubator Program that made diff_classifier
+expertise provided through the Incubator Program that made diffclassifier
 possible. The University of Washington eScience Institute is supported through a
 grant from the Gordon & Betty Moore Foundation and the Alfred P. Sloan
 Foundation. The authors would also like to thank funding from the National
