+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'diff\_classifier: Parallelization of multi-particle tracking video analyses'
+title: 'diff$\_$classifier: Parallelization of multi-particle tracking video analyses'
 tags:
 - Python
 - multi-particle tracking
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-The [diff\_classifier](https://github.com/ccurtis7/diff\_classifier) package seeks
+The [diff$\_$classifier](https://github.com/ccurtis7/diff$\_$classifier) package seeks
 to address the issue of scale-up in multi-particle tracking (MPT) analyses via a
 parallelization approach. MPT is a powerful analytical tool that has been used
 in fields ranging from aeronautics to oceanography [@Pulford:2005] allowing
@@ -53,7 +53,7 @@ towards short fast-moving trajectories and could result in empty datasets
 Due to variations in experimental conditions and image quality, user-selected
 tracking parameters can vary widely from video to video. As parameter selection
 can also vary from user to user, this also brings up the issue of
-reproducibility. diff\_classifier addresses these issues with regression
+reproducibility. diff$\_$classifier addresses these issues with regression
 tools to predict input tracking parameters and parallelized script-based
 implementations in Amazon Web Services (AWS), using the Simple Storage
 Service (S3) and Batch for data storage and computing, respectively, and
@@ -64,7 +64,7 @@ tracking parameters based on intensity distributions of input images. This can
 simultaneously reduce time-to-first-result in MPT workflows and provide
 reproducible MPT results.
 
-diff\_classifier also includes downstream MPT analysis tools including mean
+diff$\_$classifier also includes downstream MPT analysis tools including mean
 squared displacement and feature calculations, visualization tools, and a
 principle component analysis implementation. MPT is commonly used to calculate
 and report ensemble-averaged diffusion coefficients of nanoparticles and other
@@ -74,7 +74,7 @@ analysis to individual particle trajectories. By including a variety of features
 resolution, users can implement a range of data science analysis techniques to
 their MPT datasets.
 
-The source code for diff\_classifier has been archived to Zenodo with the
+The source code for diff$\_$classifier has been archived to Zenodo with the
 linked DOI: [@zenodo]
 
 
@@ -82,7 +82,7 @@ linked DOI: [@zenodo]
 
 The authors would like to thank the eScience Institute for the resources
 and expertise provided through the Incubator Program that made
-diff\_classifier possible. The University of Washington eScience Institute
+diff$\_$classifier possible. The University of Washington eScience Institute
 is supported through a grant from the Gordon & Betty Moore Foundation and
 the Alfred P. Sloan Foundation. The authors would also like to thank
 funding from the National Institute of General Medical Sciences 1R35
