+++ paper.bib
@@ -1,0 +1,56 @@
+@ARTICLE{Pulford:2005,
+  author = "Pulford G W",
+  title = "Taxonomy of multiple target tracking methods",
+  journal = "Iee Proceedings - Radar Sonar and Navigation",
+  volume = "152",
+  number = "5",
+  year   = "2005",
+  doi    = "10.1049/ip-rsn:20045064"
+}
+
+@ARTICLE{Chenouard:2014,
+  author  = "Chenouard N",
+  title   = "Objective comparison of particle tracking methods",
+  journal = "Nature Methods",
+  volume  = "11",
+  number  = "3",
+  year    = "2014",
+  doi     = "10.1038/nmeth.2808"
+}
+
+@ARTICLE{Tineves:2017,
+  author  = "Tineves J, Perry N, Schindelin J, Hoopes G M, Reynolds G D, Laplantine E, Bednarek S Y, Shorte S L, and Eliceiri K W",
+  title   = "TrackMate$\colon$ An open and extensible platform for single-particle tracking",
+  journal = "Methods",
+  volume  = "115",
+  year    = "2017",
+  doi     = "10.1016/j.ymeth.2016.09.016"
+}
+
+@ARTICLE{Wang:2015,
+  author  = "Wang Y Y, Nunn K L, Harit D, McKinley S A, and Lai S K",
+  title   = "Minimizing biases associated with tracking analysis of submicron particles in heterogeneous biological fluids",
+  journal = "Journal of Controlled Release",
+  volume  = "220",
+  number  = "PtA",
+  year    = "2015",
+  doi     = "10.1016/j.jconrel.2015.10.021"
+}
+
+@MISC{zenodo,
+  title     = "diff$\_$classifier",
+  author    = "Curtis C, Rokem A, and Nance E",
+  publisher = "Zenodo",
+  year      = "2018",
+  doi       = "10.5281/zenodo.2631862"
+}
+
+@InProceedings{cloudknot,
+  author    = "Richie-Halford A, Rokem A",
+  title     = "r Cloudknot$\colon$ A Python Library to Run your Existing Code on AWS Batch",
+  booktitle = "Proceedings of the 17th Python in Science Conference",
+  pages     = "8-14",
+  year      = "2018",
+  editor    = "Akici F, Lippa D, Niederhut D, Pacer M",
+  doi       = "10.25080/Majora-4af1f417-001"
+}
