+++ paper.bib
@@ -38,7 +38,7 @@
 }
 
 @MISC{zenodo,
-  title     = "diff$\_$classifier",
+  title     = "diffclassifier",
   author    = "Curtis C, Rokem A, and Nance E",
   publisher = "Zenodo",
   year      = "2018",
