+++ paper.bib
@@ -38,7 +38,7 @@
 }
 
 @MISC{zenodo,
-  title     = "diffclassifier",
+  title     = "diff$\_$classifier",
   author    = "Curtis C, Rokem A, and Nance E",
   publisher = "Zenodo",
   year      = "2018",
