+++ paper.bib
@@ -38,7 +38,7 @@
 }
 
 @MISC{zenodo,
-  title     = "diff$\_$classifier",
+  title     = "diff\_classifier",
   author    = "Curtis C, Rokem A, and Nance E",
   publisher = "Zenodo",
   year      = "2018",
