+++ paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Pulford:2005,
-  author = "Pulford G W",
+  author = "Graham W. Pulford",
   title = "Taxonomy of multiple target tracking methods",
   journal = "Iee Proceedings - Radar Sonar and Navigation",
   volume = "152",
@@ -9,7 +9,7 @@
 }
 
 @ARTICLE{Chenouard:2014,
-  author  = "Chenouard N",
+  author  = "Nicolas Chenouard, Ihor Smal, Fabrice de Chaumont, Martin Ma\v{s}ka, Ivo F Sbalzarini, Yuanhao Gong, Janick Cardinale, Craig Carthel, Stefano Coraluppi, Mark Winter, Andrew R. Cohen, Willian J. Godinez, Karl Rohr, Yannis Kalaidzidis, Liang Liang, James Duncan, Hongying Shen, Yingke Xu, Klas E. G. Magnusson, Joakim Jald\'{e}n, Helen M. Blau, Perrine Paul-Gilloteaux, Philippe Roudot, Charles Kervrann, Fran\c{c}ois Waharte, Jean-Yves Tinevez, Spencer L. Shorte, Joost Willemse, Katherine Celler, Gille P. van Wezel, Han-Wei Dan, Yuh-Show Tsai, Carlos Ortiz de Sol\'{o}rzano, Jean-Christophe Olivo-Marin, \& Erik Meijering",
   title   = "Objective comparison of particle tracking methods",
   journal = "Nature Methods",
   volume  = "11",
@@ -19,7 +19,7 @@
 }
 
 @ARTICLE{Tineves:2017,
-  author  = "Tineves J, Perry N, Schindelin J, Hoopes G M, Reynolds G D, Laplantine E, Bednarek S Y, Shorte S L, and Eliceiri K W",
+  author  = "Jean-Yves Tineves, Nick Perry, Johannes Schindelin, Genevieve M. Hoopes, Gregory D. Reynolds, Emmanuel Laplantine, Sebastian Y. Bednarek, Spencer L. Shorte, \& Kevin W. Eliceiri",
   title   = "TrackMate$\colon$ An open and extensible platform for single-particle tracking",
   journal = "Methods",
   volume  = "115",
@@ -28,7 +28,7 @@
 }
 
 @ARTICLE{Wang:2015,
-  author  = "Wang Y Y, Nunn K L, Harit D, McKinley S A, and Lai S K",
+  author  = "Ying-Yang Wang, Kenetta L. Nunn, Dimple Harit, Scott A. McKinley, \& Samuel K. Lai",
   title   = "Minimizing biases associated with tracking analysis of submicron particles in heterogeneous biological fluids",
   journal = "Journal of Controlled Release",
   volume  = "220",
@@ -39,15 +39,15 @@
 
 @MISC{zenodo,
   title     = "diff\_classifier",
-  author    = "Curtis C, Rokem A, and Nance E",
+  author    = "Chad D. Curtis, Ariel Rokem \& Elizabeth Nance",
   publisher = "Zenodo",
   year      = "2018",
   doi       = "10.5281/zenodo.2631862"
 }
 
 @InProceedings{cloudknot,
-  author    = "Richie-Halford A, Rokem A",
-  title     = "r Cloudknot$\colon$ A Python Library to Run your Existing Code on AWS Batch",
+  author    = "Adam Richie-Halford, \& Ariel Rokem",
+  title     = "Cloudknot\colon A Python Library to Run your Existing Code on AWS Batch",
   booktitle = "Proceedings of the 17th Python in Science Conference",
   pages     = "8-14",
   year      = "2018",
