### 2019-04-09 c110fe2
Fixed capitalization of diff_classifier in .bib file
