+++ paper.bib
@@ -38,7 +38,7 @@
 }
 
 @MISC{zenodo,
-  title     = "diff\_classifier",
+  title     = {{d}iff\_classifier},
   author    = {{C}had {D.} {C}urtis, {A}riel {R}okem \& {E}lizabeth {N}ance},
   publisher = "Zenodo",
   year      = "2018",
