### 2019-04-09 e58151d
Tried modeling author lists after Ariel's suggestion.
