+++ paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Pulford:2005,
-  author = {{{G}raham {W.} {P}ulford}},
+  author = {{G}raham {W}. {P}ulford},
   title = "Taxonomy of multiple target tracking methods",
   journal = "Iee Proceedings - Radar Sonar and Navigation",
   volume = "152",
@@ -9,7 +9,7 @@
 }
 
 @ARTICLE{Chenouard:2014,
-  author  = {{{N}icolas {C}henouard} \& others},
+  author  = {{N}icolas {C}henouard and others},
   title   = "Objective comparison of particle tracking methods",
   journal = "Nature Methods",
   volume  = "11",
@@ -19,7 +19,7 @@
 }
 
 @ARTICLE{Tineves:2017,
-  author  = {{J}ean-{Y}ves {T}ineves \& others},
+  author  = {{J}ean-{Y}ves {T}ineves and others},
   title   = "TrackMate$\colon$ An open and extensible platform for single-particle tracking",
   journal = "Methods",
   volume  = "115",
@@ -28,7 +28,7 @@
 }
 
 @ARTICLE{Wang:2015,
-  author  = {{Y}ing-{Y}ang {W}ang, {K}enetta {L.} {N}unn, {D}imple {H}arit, {S}cott {A.} {M}c{K}inley, \& {S}amuel {K.} {L}ai},
+  author  = {{Y}ing-{Y}ang {W}ang, {K}enetta {L}. {N}unn, {D}imple {H}arit, {S}cott {A}. {M}c{K}inley and {S}amuel {K}. {L}ai},
   title   = "Minimizing biases associated with tracking analysis of submicron particles in heterogeneous biological fluids",
   journal = "Journal of Controlled Release",
   volume  = "220",
@@ -39,15 +39,15 @@
 
 @MISC{zenodo,
   title     = {{d}iff\_classifier},
-  author    = {{C}had {D.} {C}urtis, {A}riel {R}okem \& {E}lizabeth {N}ance},
+  author    = {{C}had {D}. {C}urtis, {A}riel {R}okem and {E}lizabeth {N}ance},
   publisher = "Zenodo",
   year      = "2018",
   doi       = "10.5281/zenodo.2631862"
 }
 
 @InProceedings{cloudknot,
-  author    = {{{A}dam {R}ichie-{H}alford}, \& {{A}riel {R}okem}},
-  title     = "Cloudknot\colon A Python Library to Run your Existing Code on AWS Batch",
+  author    = {{A}dam {R}ichie-{H}alford and {A}riel {R}okem},
+  title     = "{C}loudknot A Python Library to Run your Existing Code on AWS Batch",
   booktitle = "Proceedings of the 17th Python in Science Conference",
   pages     = "8-14",
   year      = "2018",
