### 2019-04-09 3624d8f
Modified .bib file after a JOSS template.
