+++ paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Pulford:2005,
-  author = {{G}raham {W}. {P}ulford},
+  author = {Pulford, Graham W.},
   title = "Taxonomy of multiple target tracking methods",
   journal = "Iee Proceedings - Radar Sonar and Navigation",
   volume = "152",
@@ -9,7 +9,7 @@
 }
 
 @ARTICLE{Chenouard:2014,
-  author  = {{N}icolas {C}henouard and others},
+  author  = {Chenouard, Nicholas and Smal, Ihor and de Chaumont, Fabrice and Ma\v{s}ka, Martin and Sbalzarini, Ivo F. and Gong, Yuanhao and Cardinale, Janick and Carthel, Craig and Coraluppi, Stefano and Winter, Mark and Cohen, Andrew R. and Godinez, Willian J. and Rohr, Karl and Kalaidzidis, Yannis and Liang, Liang and Duncan, James and Shen, Hongying and Xu, Yingke and Magnusson, Klas E. G. and Jald\'{e}n, Joakim and Blau, Helen M. and Paul-Gilloteaux, Perrine and Roudot, Philippe and Kervrann, Charles and Waharte, Fran\c{c}ois and Tinevez, Jean-Yves and Shorte, Spencer L. and Willemse, Joost and Celler, Katherine and van Wezel, Gille P. and Dan, Han-Wei and Tsai, Yuh-Show and de Sol\'{o}rzano, Carlos Ortiz and Olivo-Marin, Jean-Christophe and Meijering, Erik},
   title   = "Objective comparison of particle tracking methods",
   journal = "Nature Methods",
   volume  = "11",
@@ -19,7 +19,7 @@
 }
 
 @ARTICLE{Tineves:2017,
-  author  = {{J}ean-{Y}ves {T}ineves and others},
+  author  = {Tineves, Jean-Yves and Perry, Nick and Schindelin, Johannes and Hoopes, Genevieve M. and Reynolds, Gregory D. and Laplantine, Emmanuel and Bednarek, Sebastian Y. and Shorte, Spencer L. and Eliceiri, Kevin W.},
   title   = "TrackMate$\colon$ An open and extensible platform for single-particle tracking",
   journal = "Methods",
   volume  = "115",
@@ -28,7 +28,7 @@
 }
 
 @ARTICLE{Wang:2015,
-  author  = {{Y}ing-{Y}ang {W}ang, {K}enetta {L}. {N}unn, {D}imple {H}arit, {S}cott {A}. {M}c{K}inley and {S}amuel {K}. {L}ai},
+  author  = {Wang, Ying-Yang and Nunn, Kenetta L. and Harit, Dimple and McKinley, Scott A. and Lai, Samuel K.},
   title   = "Minimizing biases associated with tracking analysis of submicron particles in heterogeneous biological fluids",
   journal = "Journal of Controlled Release",
   volume  = "220",
@@ -39,14 +39,14 @@
 
 @MISC{zenodo,
   title     = {{d}iff\_classifier},
-  author    = {{C}had {D}. {C}urtis, {A}riel {R}okem and {E}lizabeth {N}ance},
+  author    = {Curtis, Chad D. and Rokem, Ariel and Nance, Elizabeth},
   publisher = "Zenodo",
   year      = "2018",
   doi       = "10.5281/zenodo.2631862"
 }
 
 @InProceedings{cloudknot,
-  author    = {{A}dam {R}ichie-{H}alford and {A}riel {R}okem},
+  author    = {Richie-Halford, Adam and Rokem, Ariel},
   title     = "{C}loudknot A Python Library to Run your Existing Code on AWS Batch",
   booktitle = "Proceedings of the 17th Python in Science Conference",
   pages     = "8-14",
