+++ paper.bib
@@ -1,11 +1,11 @@
 @ARTICLE{Pulford:2005,
   author = {Pulford, Graham W.},
-  title = "Taxonomy of multiple target tracking methods",
-  journal = "Iee Proceedings - Radar Sonar and Navigation",
-  volume = "152",
-  number = "5",
-  year   = "2005",
-  doi    = "10.1049/ip-rsn:20045064"
+  title = {Taxonomy of multiple target tracking methods},
+  journal = {Iee Proceedings - Radar Sonar and Navigation},
+  volume = {152},
+  number = {5},
+  year   = {2005},
+  doi    = {10.1049/ip-rsn:20045064}
 }
 
 @ARTICLE{Chenouard:2014,
@@ -47,7 +47,7 @@
 
 @InProceedings{cloudknot,
   author    = {Richie-Halford, Adam and Rokem, Ariel},
-  title     = "{C}loudknot A Python Library to Run your Existing Code on AWS Batch",
+  title     = "Cloudknot: {A} Python Library to Run your Existing Code on AWS Batch",
   booktitle = "Proceedings of the 17th Python in Science Conference",
   pages     = "8-14",
   year      = "2018",
