### 2019-04-09 #989 @ccurtis7
Hmmm, it appears that whedon has changed how it treats Latex characters in the title. Previously, I had to type `diff\_classifier` for it to render correctly, but now I no longer have to warn it with a `\`. I can just type `diff_classifier`. I have updated the file in my master branch accordingly.

Do I have to generate another release, or is the `paper.md` file in the master branch OK @cMadan?
