### 2019-04-09 954e4de
Removed Zenodo ref in paper files
