+++ paper.bib
@@ -37,14 +37,6 @@
   doi     = "10.1016/j.jconrel.2015.10.021"
 }
 
-@MISC{zenodo,
-  title     = {{d}iff\_classifier},
-  author    = {Curtis, Chad D. and Rokem, Ariel and Nance, Elizabeth},
-  publisher = "Zenodo",
-  year      = "2018",
-  doi       = "10.5281/zenodo.2631862"
-}
-
 @InProceedings{cloudknot,
   author    = {Richie-Halford, Adam and Rokem, Ariel},
   title     = "Cloudknot: {A} Python Library to Run your Existing Code on AWS Batch",
