### 2019-04-09 #989 @ccurtis7
I have made the requested changes @labarba, thank you.
