### 2019-04-10 #989 @labarba
Congratulations, @ccurtis7, your JOSS paper is published!

Sincere thanks to the editor: @cMadan, and reviewer: @stsievert — Merci, gracias, danke 🙏
